for/requirements.txt
Django==4.2.7
antlr4-python3-runtime==4.12.0